branching optimization
For example, when a ray bounces of a semi reflective / semi transparent
material, it used to send off multiple rays (because it has to handle
both cases of reflecting and not reflecting). This causes an up to
O(2^n) if there's a hall of mirrors situation (which sometimes happens
with inner reflections in a crystal-like material.